refactor: sieve
